アーキテクチャに沿ったエンティティを作成
