+++ src/main/kotlin/com/example/kyudo_app/domain/model/GameMatch.kt
@@ -1,32 +1,11 @@
 package com.example.kyudo_app.domain.model
 
-import jakarta.persistence.*
-import lombok.Getter
-import lombok.Setter
-import java.util.*
 
-@Setter
-@Getter
-@Entity
-@Table(name = "game_match")
-class GameMatch {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "match_id")
-    private var matchId: Long? = null
+import java.util.Date
 
-    @Column(name = "match_date")
-    private var matchDate: Date? = null
-
-    @ManyToMany
-    @JoinTable(
-        name = "GameMatch_User",
-        joinColumns = [JoinColumn(name = "match_id")],
-        inverseJoinColumns = [JoinColumn(name = "user_id")]
-    )
-    private val users: MutableList<User?>? = null // GameMatchに関連する複数のユーザー
-
-    @ManyToOne
-    @JoinColumn(name = "Place_id", referencedColumnName = "Place_id")
-    private val place: Place? = null
-}
+class GameMatch(
+    val matchId: Long?,
+    val matchDate: Date?,
+    val users: List<User> = emptyList(),
+    val place: Place? = null
+)
