+++ src/main/kotlin/com/example/kyudo_app/domain/model/Team.kt
@@ -1,19 +1,7 @@
 package com.example.kyudo_app.domain.model
 
-import jakarta.persistence.*
-import lombok.Getter
-import lombok.Setter
 
-@Setter
-@Getter
-@Entity
-@Table(name = "Team")
-class Team {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "team_id")
-    private var teamId: Long? = null
-
-    @Column(name = "team_name")
-    private var teamName: String? = null
-}
+class Team(
+    val teamId: Long?,
+    val teamName: String?
+)
